Refactored TimestampProvider plugin
