+++ packages/live-share/src/LiveEvent.ts
@@ -236,8 +236,8 @@ export class LiveEvent<
     }
 
     /**
-     * @hidden
-     * Assigns a new timestamp provider.
+     * Assigns a custom timestamp provider.
+     * @provider 
      */
     public static setTimestampProvider(provider: ITimestampProvider): void {
         LiveEvent._timestampProvider = provider;
