+++ packages/live-share/src/internals/index.ts
@@ -8,6 +8,5 @@ export * from "./DefaultTeamsHost";
 export * from "./LiveShareTokenProvider";
 export * from "./RequestCache";
 export * from "./RoleVerifier";
-export * from "./SharedClock";
 export * from "./TeamsClientApi";
 export * from "./utils";
